refactor: extract ultraplan.go into focused subpackage
Refactors the monolithic ultraplan.go (1,903 LOC) into focused,
testable components in a new internal/tui/view/ultraplan/ subpackage.

The split follows Bubbletea best practices where views are composed
from smaller, focused rendering functions:

- context.go: State and RenderContext types
- header.go: HeaderRenderer for phase and progress display
- tasks.go: TaskRenderer for task line rendering and group stats
- status.go: StatusRenderer for phase indicators
- sidebar.go: SidebarRenderer for unified sidebar composition
- consolidation.go: ConsolidationRenderer for consolidation phase
- help.go: HelpRenderer for context-sensitive help bar
- planview.go: PlanViewRenderer for detailed plan view
- inline.go: InlineRenderer for collapsible group content
- helpers.go: Shared utilities (progress bar, truncation, etc.)
- view.go: View composition struct

The original ultraplan.go now provides backward-compatible wrappers
using type aliases and delegation methods, ensuring existing code
continues to work without modification.

Includes fixes for:
- Potential nil pointer dereferences in multiple renderers
- Removed broken PhaseToString and ComplexityIndicator methods
- Removed unused utility functions
- Fixed discarded return value in renderConsolidationSection
