+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Prompt Building Refactor** - Extracted prompt-building logic from coordinator.go into the orchestrator/prompt package, improving code organization and testability. Updated `buildPlanManagerPrompt` and `buildPlanComparisonSection` methods to use `PlanningBuilder` with the new conversion helpers. Added `FormatDetailedPlans` and `BuildCompactPlanManagerPrompt` methods to `PlanningBuilder` for flexible plan formatting. Exported `PlanManagerPromptTemplate` in the prompt package.
 - **Complete Coordinator Prompt Builder Migration** - Migrated remaining coordinator prompt-building methods (`buildRevisionPrompt`, `buildSynthesisPrompt`, `buildConsolidationPrompt`) to use the extracted prompt builders (`RevisionBuilder`, `SynthesisBuilder`, `ConsolidationBuilder`). Removed duplicate prompt templates from `ultraplan.go` that are now defined in the `orchestrator/prompt` package. Removed unused `PromptAdapter` struct and its methods (keeping only the standalone type conversion functions that are actively used). This completes the separation of prompt-building concerns from the Coordinator.
+- **Ultraplan View Package Granularity** - Refactored the 1,903-line `ultraplan.go` view file into focused, testable components in a new `internal/tui/view/ultraplan/` subpackage. Split rendering logic into separate files by concern: `header.go` (header with phase and progress), `tasks.go` (task list rendering with status icons and wrapping), `status.go` (phase indicators and progress bars), `consolidation.go` (consolidation phase sidebar), `help.go` (context-sensitive help bar), `sidebar.go` (unified sidebar composition), `planview.go` (detailed plan view), and `inline.go` (inline content for collapsible groups). The original `ultraplan.go` now provides backward-compatible wrappers that delegate to the subpackage. This follows Bubbletea best practices where views are composed from smaller, focused rendering functions.
 
 ### Fixed
