+++ internal/tui/view/ultraplan/consolidation.go
@@ -1,0 +1,174 @@
+package ultraplan
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/Iron-Ham/claudio/internal/orchestrator"
+	"github.com/Iron-Ham/claudio/internal/tui/styles"
+	"github.com/charmbracelet/lipgloss"
+)
+
+// ConsolidationRenderer handles rendering of the consolidation phase sidebar.
+type ConsolidationRenderer struct {
+	ctx *RenderContext
+}
+
+// NewConsolidationRenderer creates a new consolidation renderer with the given context.
+func NewConsolidationRenderer(ctx *RenderContext) *ConsolidationRenderer {
+	return &ConsolidationRenderer{ctx: ctx}
+}
+
+// RenderSidebar renders the sidebar during the consolidation phase.
+func (c *ConsolidationRenderer) RenderSidebar(width int, height int) string {
+	if c.ctx.UltraPlan == nil || c.ctx.UltraPlan.Coordinator == nil {
+		return ""
+	}
+	session := c.ctx.UltraPlan.Coordinator.Session()
+	if session == nil || session.Consolidation == nil {
+		return ""
+	}
+
+	var b strings.Builder
+	state := session.Consolidation
+
+	// Title with phase indicator
+	b.WriteString(styles.SidebarTitle.Render("Consolidation"))
+	b.WriteString("\n\n")
+
+	// Phase status
+	phaseIcon := ConsolidationPhaseIcon(state.Phase)
+	phaseDesc := ConsolidationPhaseDesc(state.Phase)
+	statusLine := fmt.Sprintf("%s %s", phaseIcon, phaseDesc)
+	b.WriteString(statusLine)
+	b.WriteString("\n\n")
+
+	// Progress: Groups
+	if state.TotalGroups > 0 {
+		groupProgress := fmt.Sprintf("Groups: %d/%d", state.CurrentGroup, state.TotalGroups)
+		b.WriteString(styles.Muted.Render(groupProgress))
+		b.WriteString("\n")
+
+		for i, branch := range state.GroupBranches {
+			prefix := "  ✓ "
+			if i == state.CurrentGroup-1 && state.Phase != orchestrator.ConsolidationComplete {
+				prefix = "  ⟳ "
+			}
+			branchDisplay := truncate(branch, width-8)
+			b.WriteString(styles.Muted.Render(prefix + branchDisplay))
+			b.WriteString("\n")
+		}
+		b.WriteString("\n")
+	}
+
+	// Current task being merged
+	if state.CurrentTask != "" {
+		b.WriteString(styles.Muted.Render("Merging:"))
+		b.WriteString("\n")
+		taskDisplay := truncate(state.CurrentTask, width-6)
+		b.WriteString(styles.Muted.Render("  " + taskDisplay))
+		b.WriteString("\n\n")
+	}
+
+	// PRs created
+	if len(state.PRUrls) > 0 {
+		b.WriteString(styles.SidebarTitle.Render("Pull Requests"))
+		b.WriteString("\n")
+		for i, prURL := range state.PRUrls {
+			prefix := "  ✓ "
+			prDisplay := fmt.Sprintf("PR #%d", i+1)
+			if idx := strings.LastIndex(prURL, "/"); idx >= 0 {
+				prDisplay = "PR " + prURL[idx+1:]
+			}
+			b.WriteString(styles.Muted.Render(prefix + prDisplay))
+			b.WriteString("\n")
+		}
+		b.WriteString("\n")
+	}
+
+	// Conflict info (if paused)
+	if state.Phase == orchestrator.ConsolidationPaused && len(state.ConflictFiles) > 0 {
+		b.WriteString(lipgloss.NewStyle().Foreground(styles.RedColor).Render("⚠ Conflict Detected"))
+		b.WriteString("\n\n")
+		b.WriteString(styles.Muted.Render("Files:"))
+		b.WriteString("\n")
+		maxFiles := 5
+		for i, file := range state.ConflictFiles {
+			if i >= maxFiles {
+				remaining := len(state.ConflictFiles) - maxFiles
+				b.WriteString(styles.Muted.Render(fmt.Sprintf("  ... +%d more", remaining)))
+				b.WriteString("\n")
+				break
+			}
+			fileDisplay := truncate(file, width-6)
+			b.WriteString(styles.Muted.Render("  " + fileDisplay))
+			b.WriteString("\n")
+		}
+		b.WriteString("\n")
+		b.WriteString(styles.Muted.Render("Press [r] to resume"))
+		b.WriteString("\n")
+	}
+
+	// Error message
+	if state.Error != "" {
+		b.WriteString(lipgloss.NewStyle().Foreground(styles.RedColor).Render("Error:"))
+		b.WriteString("\n")
+		errDisplay := truncate(state.Error, width-4)
+		b.WriteString(styles.Muted.Render(errDisplay))
+		b.WriteString("\n")
+	}
+
+	return styles.Sidebar.Width(width - 2).Render(b.String())
+}
+
+// ConsolidationPhaseIcon returns an icon for the consolidation phase.
+func ConsolidationPhaseIcon(phase orchestrator.ConsolidationPhase) string {
+	switch phase {
+	case orchestrator.ConsolidationIdle:
+		return "○"
+	case orchestrator.ConsolidationDetecting:
+		return "⟳"
+	case orchestrator.ConsolidationCreatingBranches:
+		return "⟳"
+	case orchestrator.ConsolidationMergingTasks:
+		return "⟳"
+	case orchestrator.ConsolidationPushing:
+		return "⟳"
+	case orchestrator.ConsolidationCreatingPRs:
+		return "⟳"
+	case orchestrator.ConsolidationPaused:
+		return "⏸"
+	case orchestrator.ConsolidationComplete:
+		return "✓"
+	case orchestrator.ConsolidationFailed:
+		return "✗"
+	default:
+		return "○"
+	}
+}
+
+// ConsolidationPhaseDesc returns a human-readable description for the consolidation phase.
+func ConsolidationPhaseDesc(phase orchestrator.ConsolidationPhase) string {
+	switch phase {
+	case orchestrator.ConsolidationIdle:
+		return "Waiting..."
+	case orchestrator.ConsolidationDetecting:
+		return "Detecting conflicts..."
+	case orchestrator.ConsolidationCreatingBranches:
+		return "Creating branches..."
+	case orchestrator.ConsolidationMergingTasks:
+		return "Merging tasks..."
+	case orchestrator.ConsolidationPushing:
+		return "Pushing to remote..."
+	case orchestrator.ConsolidationCreatingPRs:
+		return "Creating PRs..."
+	case orchestrator.ConsolidationPaused:
+		return "Paused (conflict)"
+	case orchestrator.ConsolidationComplete:
+		return "Complete"
+	case orchestrator.ConsolidationFailed:
+		return "Failed"
+	default:
+		return string(phase)
+	}
+}
