+++ internal/tui/view/ultraplan/status.go
@@ -1,0 +1,191 @@
+package ultraplan
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/Iron-Ham/claudio/internal/orchestrator"
+	"github.com/Iron-Ham/claudio/internal/tui/styles"
+	"github.com/charmbracelet/lipgloss"
+)
+
+// StatusRenderer handles rendering of phase status indicators and progress displays.
+type StatusRenderer struct {
+	ctx *RenderContext
+}
+
+// NewStatusRenderer creates a new status renderer with the given context.
+func NewStatusRenderer(ctx *RenderContext) *StatusRenderer {
+	return &StatusRenderer{ctx: ctx}
+}
+
+// RenderProgressBar renders a simple ASCII progress bar.
+func RenderProgressBar(percent int, width int) string {
+	if percent < 0 {
+		percent = 0
+	}
+	if percent > 100 {
+		percent = 100
+	}
+
+	filled := (percent * width) / 100
+	empty := width - filled
+
+	bar := strings.Repeat("█", filled) + strings.Repeat("░", empty)
+	return fmt.Sprintf("[%s]", bar)
+}
+
+// PhaseToString converts a phase to a display string.
+func PhaseToString(phase orchestrator.UltraPlanPhase) string {
+	switch phase {
+	case orchestrator.PhasePlanning:
+		return "PLANNING"
+	case orchestrator.PhasePlanSelection:
+		return "SELECTING PLAN"
+	case orchestrator.PhaseRefresh:
+		return "READY"
+	case orchestrator.PhaseExecuting:
+		return "EXECUTING"
+	case orchestrator.PhaseSynthesis:
+		return "SYNTHESIS"
+	case orchestrator.PhaseRevision:
+		return "REVISION"
+	case orchestrator.PhaseConsolidating:
+		return "CONSOLIDATING"
+	case orchestrator.PhaseComplete:
+		return "COMPLETE"
+	case orchestrator.PhaseFailed:
+		return "FAILED"
+	default:
+		return string(phase)
+	}
+}
+
+// PhaseStyle returns the style for a phase indicator.
+func PhaseStyle(phase orchestrator.UltraPlanPhase) lipgloss.Style {
+	switch phase {
+	case orchestrator.PhasePlanning:
+		return lipgloss.NewStyle().Foreground(styles.BlueColor)
+	case orchestrator.PhasePlanSelection:
+		return lipgloss.NewStyle().Foreground(lipgloss.Color("208")).Bold(true)
+	case orchestrator.PhaseRefresh:
+		return lipgloss.NewStyle().Foreground(styles.YellowColor)
+	case orchestrator.PhaseExecuting:
+		return lipgloss.NewStyle().Foreground(styles.BlueColor).Bold(true)
+	case orchestrator.PhaseSynthesis:
+		return lipgloss.NewStyle().Foreground(styles.PurpleColor)
+	case orchestrator.PhaseRevision:
+		return lipgloss.NewStyle().Foreground(lipgloss.Color("208")).Bold(true)
+	case orchestrator.PhaseConsolidating:
+		return lipgloss.NewStyle().Foreground(styles.YellowColor).Bold(true)
+	case orchestrator.PhaseComplete:
+		return lipgloss.NewStyle().Foreground(styles.GreenColor)
+	case orchestrator.PhaseFailed:
+		return lipgloss.NewStyle().Foreground(styles.RedColor)
+	default:
+		return lipgloss.NewStyle()
+	}
+}
+
+// ComplexityIndicator returns a visual indicator for task complexity.
+func ComplexityIndicator(complexity orchestrator.TaskComplexity) string {
+	switch complexity {
+	case orchestrator.ComplexityLow:
+		return "◦"
+	case orchestrator.ComplexityMedium:
+		return "◎"
+	case orchestrator.ComplexityHigh:
+		return "●"
+	default:
+		return "○"
+	}
+}
+
+// GetPhaseSectionStatus returns a status indicator for a phase section header.
+func (s *StatusRenderer) GetPhaseSectionStatus(phase orchestrator.UltraPlanPhase, session *orchestrator.UltraPlanSession) string {
+	switch phase {
+	case orchestrator.PhasePlanning:
+		if session.Phase == orchestrator.PhasePlanning {
+			return "[⟳]"
+		}
+		if session.Phase == orchestrator.PhasePlanSelection {
+			return "[✓]"
+		}
+		if session.Plan != nil {
+			return "[✓]"
+		}
+		return "[○]"
+
+	case orchestrator.PhasePlanSelection:
+		if session.Phase == orchestrator.PhasePlanSelection {
+			return "[⟳]"
+		}
+		if session.Phase == orchestrator.PhaseRefresh ||
+			session.Phase == orchestrator.PhaseExecuting ||
+			session.Phase == orchestrator.PhaseSynthesis ||
+			session.Phase == orchestrator.PhaseRevision ||
+			session.Phase == orchestrator.PhaseConsolidating ||
+			session.Phase == orchestrator.PhaseComplete {
+			return "[✓]"
+		}
+		return "[○]"
+
+	case orchestrator.PhaseExecuting:
+		if session.Plan == nil {
+			return "[○]"
+		}
+		completed := len(session.CompletedTasks)
+		total := len(session.Plan.Tasks)
+		if session.Phase == orchestrator.PhaseExecuting {
+			return fmt.Sprintf("[%d/%d]", completed, total)
+		}
+		if completed == total && total > 0 {
+			return "[✓]"
+		}
+		if completed > 0 {
+			return fmt.Sprintf("[%d/%d]", completed, total)
+		}
+		return "[○]"
+
+	case orchestrator.PhaseSynthesis:
+		if session.Phase == orchestrator.PhaseSynthesis {
+			return "[⟳]"
+		}
+		if session.Phase == orchestrator.PhaseRevision || session.Phase == orchestrator.PhaseConsolidating || session.Phase == orchestrator.PhaseComplete {
+			return "[✓]"
+		}
+		return "[○]"
+
+	case orchestrator.PhaseRevision:
+		if session.Phase == orchestrator.PhaseRevision {
+			if session.Revision != nil {
+				revised := len(session.Revision.RevisedTasks)
+				total := len(session.Revision.TasksToRevise)
+				return fmt.Sprintf("[%d/%d]", revised, total)
+			}
+			return "[⟳]"
+		}
+		if session.Phase == orchestrator.PhaseConsolidating || session.Phase == orchestrator.PhaseComplete {
+			if session.Revision != nil && len(session.Revision.Issues) > 0 {
+				return "[✓]"
+			}
+			return "[○]"
+		}
+		return "[○]"
+
+	case orchestrator.PhaseConsolidating:
+		if session.Phase == orchestrator.PhaseConsolidating {
+			return "[⟳]"
+		}
+		if session.Phase == orchestrator.PhaseComplete && len(session.PRUrls) > 0 {
+			return "[✓]"
+		}
+		if session.Phase == orchestrator.PhaseComplete {
+			return "[✓]"
+		}
+		return "[○]"
+
+	default:
+		return ""
+	}
+}
